Revert "Plan 9 from Bell Labs Fourth Edition 2005-06-21"
This reverts commit e8ece30e958f586a4dcfe69d224c5cce2445b17b.

+++ sys/src/libc/port/qsort.c
@@ -2,8 +2,6 @@
  * qsort -- simple quicksort
  */
 
-#include <u.h>
-
 typedef
 struct
 {
@@ -118,7 +116,7 @@ qsort(void *va, long n, long es, int (*cmp)(void*, void*))
 	s.cmp = cmp;
 	s.es = es;
 	s.swap = swapi;
-	if(((uintptr)va | es) % sizeof(long))
+	if(((long)va | es) % sizeof(long))
 		s.swap = swapb;
 	qsorts((char*)va, n, &s);
 }
